# resources/skg-if-openapi.yaml
openapi: 3.1.0
info:
  version: 1.0.0-alpha
  title: SKG-IF OpenAPI
  termsOfService: 'https://skg-if.github.io/interoperability-framework/'
  contact:
    url: 'https://github.com/skg-if'
  license:
    name: CC-BY
    url: 'http://creativecommons.org/licenses/by/4.0/'
  x-logo:
    url: 'https://www.rd-alliance.org/wp-content/uploads/2024/04/RDA_Logotype_CMYK.png'
  description: |
    OpenAPI end points specifications for [SKG-IF](https://skg-if.github.io/interoperability-framework/)

    Last version of the SKG-IF OpenAPI is available at : https://skg-if.github.io/api/

    # Note for implementers :
      - Please refer to [SKG-IF OpenAPI Implementer documentation](https://docs.google.com/document/d/1t7b7h28UTtM56Sda4NGJIp0hnQfGbcVVGn12fny9wfI/edit?usp=sharing)
      - [RDA SKG-IF WG](https://skg-if.github.io/) does not provide any official implementation of the current specification.
      - Recommended OpenAPI viewer :
        - StopLight/Elements Viewer [Current/last SKG-IF OpenAPI version](https://elements-demo.stoplight.io/?spec=https://w3id.org/skg-if/api/skg-if-openapi.yaml).
        - Swagger : [Current/last SKG-IF OpenAPI version](https://editor-next.swagger.io/)
...
servers:
  - url: '{protocol}://{server}:{port}/{skg_if_api_path}'
    variables:
      protocol:
        enum:
          - "https"
          - "http"
        default: "https"
        description: service provider protocol.
      server:
        default: example.com
...
      port:
        default: "8443"
...
paths:
...
  '/persons/{short_local_identifier}':
    get:
      tags:
          - Person
      summary: Get person by id
      description: |
          Get `person` by id. See definition in SKG-IF [Agent](https://skg-if.github.io/interoperability-framework/docs/agent.html) ( entity_type:person ) .\
      operationId: getPersonById
      parameters:
        - $ref : '#/components/parameters/shortLocalIdPathParam'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                required: ['@context','@graph']
                properties:
                  "@context":
                    type: array
                    description: JSON-LD context # not displayed in Stoplight elements
                    minItems: 3 # http://../skg-if.json, http://../skg-if-api.json and @base
                    uniqueItems: true
                    items:
                      anyOf:
                        - $ref: '#/components/schemas/JsonLdCtxDataModel'
                        - $ref: '#/components/schemas/JsonLdCtxSearch'
                        - $ref: '#/components/schemas/JsonLdCtxBaseOrMore'
                    examples:
                      - ["https://w3id.org/skg-if/context/1.1.0/skg-if.json", "https://w3id.org/skg-if/context/1.0.0/skg-if-api.json", {"@base" : "https://w3id.org/skg-if/sandbox/my-skg-acronym/"}]
                  "@graph":
                    type: array
                    minItems: 1
                    maxItems: 1
                    items:
                      $ref: "#/components/schemas/Person"
              examples:
                PersonEx01:
                  summary: Josiah Carberry person
                  value :
                    "@context": [
                      "https://w3id.org/skg-if/context/1.1.0/skg-if.json",
                      "https://w3id.org/skg-if/context/1.0.0/skg-if-api.json",
                      {"@base" : "http://example.com/"}
                    ]
                    "@graph" :
                      - local_identifier: http://example.com/skg-if/api/persons/pers-c66c6-38be-4d5f-85db-d44c9f869333
                        entity_type: "person"
                        identifiers:
                          - scheme: orcid
                            value: 0000-0002-1825-0097
                        given_name: "Josiah"
                        family_name: "Carberry"
                        name: "Josiah Carberry"
                PersonEx02:
                  summary: John Doe person on-the-fly id
                  value :
                    "@context": [
                      "https://w3id.org/skg-if/context/1.1.0/skg-if.json",
                      "https://w3id.org/skg-if/context/1.0.0/skg-if-api.json",
                      {"@base" : "http://example.com/"}
                    ]
                    "@graph" :
                      - local_identifier: http://example.com/skg-if/api/persons/otf_126848135_prod-1-pers-1
                        entity_type: "person"
                        name: "John Doe"
        '404':
          description: |
            Error if entity does not exists
  '/persons':
    get:
      tags:
          - Person
      summary: Get list of person
      operationId: getPersons
      description: |
          Get a list of `person`. See definition in SKG-IF [Agent](https://skg-if.github.io/interoperability-framework/docs/agent.html) ( entity_type:person ) .

          __Attribute filters__ \
          You can filter using these attributes of the `person` object
          | Filter key| Attribute filter example |
          | -------- | -------- |
          | identifiers.id | `identifiers.id:0000-0002-1825-0097` |
          | identifiers.scheme | `identifiers.scheme:orcid` |
          | given_name | `given_name:Josiah` |
          | family_name |  `family_name:Carberry` |
          | name |  `name:Josiah Carberro` |
          | affiliations.affiliation.local_identifier | `affiliations.affiliation.local_identifier:http://example.com/skg-if/api/organisations/org-c66c6-38be-4d5f-85db-d44c9f869333` |
          | affiliations.affiliation.name | `affiliations.affiliation.name:Brown University` |
          | affiliations.affiliation.short_name |  `affiliations.affiliation.short_name:BU` |
          | affiliations.role | `affiliations.role:affiliate` |

          __Convenience filters__ \
          These filters aren't attributes of the `person` object, but they're handy for solving some common use cases.

          | Filter key | Filter value | List returned | Attribute filter example |
          | -------- | -------- | ------- | ------- |
          | cf.search.family_name | a string | `person` with family_name containing the filter value | `cf.search.family_name:Carberry` |
          | cf.search.given_name | a string | `person` with given_name containing the filter value | `cf.search.given_name:Jos` |
          | cf.search.name | a string | `person` with name containing the filter value | `cf.search.name:Carberry` |

          __Query parameter syntax examples__ \

          * `/persons?filter=identifiers.scheme:orcid,identifiers.id:0000-0002-1825-0097`
          * `/persons?filter=cf.search.family_name:Carberry,cf.search.given_name:Jos`
          * `/persons?filter=cf.search.family_name:Carberry`
          * `/persons?filter=cf.search.family_name:Carberry&page=1&page_size=5`

      parameters:
        - name: filter
          in: query
          description: |
                Search filter. Format : Coma separated filter_name:filter_value elements ( filter_name_1:filter_value_1,filter_name_2:filter_value_2,filter_name_3:filter_value_3... ). Server side operator used is _AND_.
          schema:
            type: string
            pattern: '^(,?.+:.+)*$'
          examples :
            orcid_ex:
              value: identifiers.scheme:orcid,identifiers.id:0000-0002-1825-0097
              summary: Get all person with identifier scheme 'orcid' and id '0000-0002-1825-0097'
        - $ref : '#/components/parameters/pageQueryParam'
        - $ref : '#/components/parameters/pageSizeQueryParam'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                properties:
                  "@context":
                    type: array
                    description: JSON-LD context # not displayed in Stoplight elements
                    minItems: 3 # http://../skg-if.json, http://../skg-if-api.json and @base
                    uniqueItems: true
                    items:
                      anyOf:
                        - $ref: '#/components/schemas/JsonLdCtxDataModel'
                        - $ref: '#/components/schemas/JsonLdCtxSearch'
                        - $ref: '#/components/schemas/JsonLdCtxBaseOrMore'
                    examples:
                      - ["https://w3id.org/skg-if/context/1.1.0/skg-if.json", "https://w3id.org/skg-if/context/1.0.0/skg-if-api.json", {"@base" : "https://w3id.org/skg-if/sandbox/my-skg-acronym/"}]
                  meta:
                    type: object
                    $ref: '#/components/schemas/Meta'
                  "@graph":
                    type: array
                    items:
                      $ref: '#/components/schemas/Person'
                required: [ "@context", "meta", "@graph" ]
              examples:
                PersonEx01:
                    summary: a list of persons
                    value :
                      "@context":
                       - "https://w3id.org/skg-if/context/1.1.0/skg-if.json"
                       - "https://w3id.org/skg-if/context/1.0.0/skg-if-api.json"
                       - "@base" : "https://w3id.org/skg-if/sandbox/my-skg-acronym/"
                      meta:
                          local_identifier: https://example.com/skg-if/api/persons?filter=family_name:Carberry&page=1
                          entity_type: search_result_page
                          next_page:
                            local_identifier: https://example.com/skg-if/api/persons?filter=family_name:Carberry&page=2
                            entity_type: search_result_page
                          part_of:
                            local_identifier: https://example.com/skg-if/api/persons?filter=family_name:Carberry
                            entity_type: search_result
                            total_items: 4
                      "@graph":
                        - local_identifier: http://example.com/skg-if/api/persons/pers-c66c6-38be-4d5f-85db-d44c9f869333
                          entity_type: "person"
                          identifiers:
                            - scheme: orcid
                              value: 0000-0002-1825-0097
                          given_name: "Josiah"
                          family_name:  "Carberry"
                          name: "Josiah Carberry"
                        - local_identifier: http://example.com/skg-if/api/persons/pers-c66c6-38be-4d5f-85db-d44c9f869888
                          entity_type: "person"
                          given_name: "Gerard"
                          family_name:  "Carberry"
                          name: "Gerard Carberry"
                #PersonEx02: # commented KO on StopLight UI, OK on Redoc UI, Not used in PRISM
                #  summary: "a list of persons - external URL"
                #  externalValue: "https://raw.githubusercontent.com/skg-if/examples/refs/heads/main/OpenCitations/oc_1.json"
          links:
            address:
              operationId: getPersonById
              parameters:
                local_identifier: $response.body#/@graph/local_identifier
...
components:
  securitySchemes:
    main_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: 'http://example.com/api/oauth/dialog'
          scopes:
            'read:users': read users info
            'write:users': modify or remove users
    api_key:
      type: apiKey
      in: header
      name: api_key
    basic_auth:
      type: http
      scheme: basic
  parameters:
    shortLocalIdPathParam:
      name: short_local_identifier
      in: path
      description: |
        entity id. URL suffix used in OpenAPI get by id operations. \
        `https://[your-server-root-skg-if-url]/[products|persons|organisations|datasources|venues|topics]/{short_local_identifier}` \
        example `http://example.com/skg-if/api/products/c66c6-38be-4d5f-85db-d44c9`
      required: true
      schema:
        type: string
#    filterQueryParam: # filterQueryParam duplicated in each search endpoint ( only way to add distinct examples for each endpoint )
#      name: filter
#      in: query
#      description: |
#            Search filter. Format : Coma separated filter_name:filter_value elements ( filter_name_1:filter_value_1,filter_name_2:filter_value_2,filter_name_3:filter_value_3... ). Server side operator used is _AND_.
#      schema:
#        type: string
#        pattern: '^(,?.+:.+)*$'
    pageQueryParam:
      name: page
      in: query
      description: |
            Search page reference (page number or search iteration token).\
            Note: As a client application you should build your iterating logic using the `meta.next_page.local_identifier` and `meta.prev_page.local_identifier` URLs
      schema:
        type: string
    pageSizeQueryParam:
      name: page_size
      in: query
      description: |
            Search page size. limit search result size.
      schema:
        type: integer
  schemas:
...
    Organisation:
      type: object
      title: "Organisation"
      description:  "`organisation` object.  See definition in SKG-IF [Agent](https://skg-if.github.io/interoperability-framework/docs/agent.html) ( entity_type:organisation )."
      allOf:
        - $ref: "#/components/schemas/Entity"
        - type: object
          required: [
            "local_identifier","entity_type"
          ]
          properties :
            entity_type:
              enum:
                - "organisation"
              type: string
              x-faker:
                fake: ["organisation"]
            identifiers:
              description: Objects representing external identifiers for the entity. See [Agent](https://skg-if.github.io/interoperability-framework/docs/agent.html#identifiers) documentation for supported schemes
              type: array
              items:
                required: [ "scheme", "value" ]
                properties:
                  scheme:
                    description: The scheme for the external identifier.
                    type: string
                    x-faker:
                      fake: ["ror"]
                    examples:
                      - 'ror'
                  value:
                    type: string
                    description: "The external identifier."
                    x-faker:
                      fake : ['{{random.alphaNumeric(10)}}']
                    examples:
                      - "https://ror.org/05gq02987"
            name:
              type: string
              x-faker: company.companyName
              examples:
                - "Brown University"
            other_names:
              description:  "Other names, perhaps in different languages, identifying an `organisation`."
              type: array
              items:
                type: string
              examples :
                - ["Brownie"]
            short_name:
              type: string
              #x-faker: company.bsAdjective
              x-faker:
                random.alpha: [{count: 4, upcase: true}]
              examples:
                - "BU"
            website:
              type: string
              description: The website URL for an Organisation
              examples:
                - "https://www.brown.edu/"
            country:
              type: string
              description: The country code expressed as [ISO 3166-1 alpha-2](https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2).
              pattern: '^[A-Z]{2}$'
              examples:
                - "US"
            types:
              description:  "The types of an `organisation`."
              type: array
              items:
                type: string
                enum:
                  - archive
                  - company
                  - education
                  - facility
                  - government
                  - healthcare
                  - nonprofit
                  - funder
                  - research
                  - unspecified
      examples:
        - local_identifier: 'otf_session12324_org_1124'
          entity_type: 'organisation'
          identifiers:
            - "scheme": ror
              "value": https://ror.org/05gq02987
          name: "Brown University"
          short_name: BU
          country: US
          types:
            - education
        - local_identifier: http://example.com/skg-if/api/organisations/otf_session12325_org_1125
          entity_type: 'organisation'
          identifiers:
            - scheme: ror
              value: https://ror.org/05gq02987
            - scheme: ror
              value: https://ror.org/0274ane14
          name: John Carter Brown Library, Brown University, Providence RI, US
          short_name: BU
          country: US
          types:
            - archive
        - local_identifier: http://example.com/skg-if/api/organisations/otf_session12325_org_1126
          entity_type: 'organisation'
          name: Lab of science, unknown University, Antarctica
        - local_identifier: http://example.com/skg-if/api/organisations/otf_session12324_org_1127
          entity_type: 'organisation'
          identifiers:
            - scheme: ror
              value: https://ror.org/00k4n6c32
          name: "European Commission"
          short_name: EU
          types:
            - funder
    Person:
      type: object
      title: "Person"
      description:  "`person` object.  See definition in SKG-IF [Agent](https://skg-if.github.io/interoperability-framework/docs/agent.html) ( entity_type:person )"
      allOf:
        - $ref: "#/components/schemas/PersonLite"
        - type: object
          required: [
            "local_identifier","entity_type"
          ]
          properties :
            affiliations:
              description : "All the affiliations of a `person`. (curriculum vitae list )"
              type: array
              items:
                required: [ "affiliation"]
                properties:
                  affiliation:
                    description : "`organisation` a `person` is affiliated with"
                    $ref: "#/components/schemas/Organisation"
                  role :
                    type: string
                    default: "affiliate"
                    x-faker:
                      fake : ['affiliate']
                  period :
                    type: object
                    description : "The time period where the Person was affiliated with an Organisation"
                    properties :
                      start:
                        type: string
                        description: start date. Format [ISO 8601](https://en.wikipedia.org/wiki/ISO_8601) compliant string ( "2024-10-01", "2024-10", "2024" ).'
                      end:
                        type: string
                        description: end date. Format [ISO 8601](https://en.wikipedia.org/wiki/ISO_8601) compliant string ( "2024-10-01", "2024-10", "2024" ).'
      examples:
        - local_identifier: http://example.com/skg-if/api/persons/otf_session12324_person_1124
          entity_type: 'person'
          identifiers:
            - scheme: orcid
              value: 0000-0002-1825-0097
          given_name: Josiah
          family_name: Carberry
          name: Josiah Carberry
          affiliations:
            role: affiliate
            period:
              start: "2010-08-24"
              end: "2020-08-24"
            affiliation:
              local_identifier: http://example.com/skg-if/api/organisatios/otf_session12324_org_1124
              entity_type: 'organisation'
              identifiers:
                - scheme: ror
                  value: https://ror.org/05gq02987
              name: Brown University
              short_name: BU
              country: US
    PersonLite:
      type: object
      title: "PersonLite"
      description:  "`person` object containing a limited set of fields.  See definition in SKG-IF [Agent](https://skg-if.github.io/interoperability-framework/docs/agent.html) ( entity_type:person )."
      allOf:
        - $ref: "#/components/schemas/Entity"
        - type: object
          required: [
            "local_identifier","entity_type"
          ]
          properties :
            entity_type:
              type: string
              enum:
                - "person"
              x-faker:
                fake: ["person"]
            identifiers:
              description: Objects representing external identifiers for the entity. See [Agent](https://skg-if.github.io/interoperability-framework/docs/agent.html#identifiers) documentation for supported schemes
              type: array
              items:
                required: [ "scheme", "value" ]
                properties:
                  scheme:
                    type: string
                    description: The scheme for the external identifier.
                    x-faker:
                      fake: ["orcid"]
                    examples :
                      - orcid
                  value:
                    type: string
                    description: "The external identifier."
                    x-faker:
                      helpers.regexpStyleStringParse: ['[1000-9999]-[1000-9999]-[1000-9999]-[1000-9999]']
                    examples :
                      - "0000-0002-1825-0097"
            given_name:
              type: string
              x-faker: name.firstName
              examples :
                - Josiah
            family_name:
              type: string
              x-faker: name.lastName
              examples :
                - Carberry
            name:
              type: string
              x-faker: name.findName
              examples :
                - Josiah Carberro
      examples:
        - local_identifier: http://example.com/skg-if/api/persons/otf_session12324_person_1124
          entity_type: 'person'
          identifiers:
            - scheme: orcid
              value: 0000-0002-1825-0097
          given_name: Josiah
          family_name: Josiah
          name: Josiah Carberry
...
    Meta:
      type: object
      required: [ "local_identifier", "entity_type" ]
      properties:
        local_identifier:
          type: string
          description: search result page id
          examples:
            - https://example.com/skg-if/api/products?filter=xxx&page=y
            - https://example.com/skg-if/api/persons?filter=xxx&page=y
        entity_type:
          type: string
          description: search result page type
          enum:
            - search_result_page
        prev_page:
          description: Previous page reference
          type: object
          $ref: '#/components/schemas/SearchResultPage'
        next_page:
          description: Next page reference
          type: object
          $ref: '#/components/schemas/SearchResultPage'
        part_of:
          description: Reference for the search result as a whole (all pages)
          type: object
          required: [ "local_identifier", "entity_type" ]
          properties:
            local_identifier:
              type: string
              description: search result id as URL
              examples:
                - https://example.com/skg-if/api/products?filter=xxx
                - https://example.com/skg-if/api/persons?filter=xxx
            entity_type:
              type: string
              description: search type. Value must be `search_result`
              enum:
                - search_result
            total_items:
              description: total number of item entities for the search
              type: integer
              x-faker:
                datatype.number: [{min: 10, max: 100}]
            first_page:
              description: First page reference of the search
              type: object
              $ref: '#/components/schemas/SearchResultPage'
            last_page:
              description: Last page reference of the search
              type: object
              $ref: '#/components/schemas/SearchResultPage'
...
    JsonLdCtxDataModel:
      type: string
      description: "URL to SKG-IF data model context supported by the current OpenAPI `https://w3id.org/skg-if/context/1.1.0/skg-if.json`"
      enum:
        - https://w3id.org/skg-if/context/1.1.0/skg-if.json
      x-faker:
        fake: ['https://w3id.org/skg-if/context/1.1.0/skg-if.json']
    JsonLdCtxSearch:
      type: string
      description: "URL to SKG-IF API context (meta section) supported by the current OpenAPI `https://w3id.org/skg-if/context/1.0.0/skg-if-api.json`"
      enum:
        - https://w3id.org/skg-if/context/1.0.0/skg-if-api.json
      x-faker:
        fake: ['https://w3id.org/skg-if/context/1.0.0/skg-if-api.json']
    JsonLdCtxBaseOrMore:
      description : 'Additional context mappings ex: `{ "@base":"https://w3id.org/skg-if/sandbox/my-skg-acronym" }`'
      properties:
        "@base":
          type: string
          description : |
            Default base URL specific to the implementer.\
            common format : `https://w3id.org/skg-if/sandbox/{my-skg-acronym}` \
            example : `https://w3id.org/skg-if/sandbox/openaire/`
          x-faker:
            fake: ['https://w3id.org/skg-if/sandbox/{{random.alpha(5)}}/']
          examples:
            - "https://w3id.org/skg-if/sandbox/openaire/"
            - "https://w3id.org/skg-if/sandbox/cessda/"
            - "https://w3id.org/skg-if/sandbox/foobar/"
      additionalProperties:
        type: string
        description: "context URL of extension entities"
        x-faker:
          fake : ['http://www.example.com/skg-if/{{random.alpha(3)}}']